Refactor validation and type-checking utilities.
Updated function signatures with type hints and improved docstrings for clarity. Adjusted imports to align with the updated module structure and enhance maintainability.
