+++ tests/test_error_handling_extended.py
@@ -8,9 +8,9 @@ from pytorch_tabnet.tab_model import TabNetClassifier
 from pytorch_tabnet.utils import check_embedding_parameters, check_input, filter_weights, infer_output_dim
 from pytorch_tabnet.utils.is_multilabel import is_multilabel
 from pytorch_tabnet.utils.label_processing import unique_labels
+from pytorch_tabnet.utils.multiclass_validation import assert_all_finite
 from pytorch_tabnet.utils.type_detection import check_classification_targets
 from pytorch_tabnet.utils.type_of_target import type_of_target
-from pytorch_tabnet.utils.validation_multi import assert_all_finite
 
 
 class TestErrorHandling(unittest.TestCase):
